Add files via upload
The ZIP file contains all the necessary Python code and data required to run the Bayesian Optimization for reaction yield optimization in organic synthesis. To use the files, simply extract the ZIP archive and follow the instructions provided in the repository.
Additionally, a scientific report is included to provide in-depth information about the project, detailing the methodology, comparison with Multiple Linear Regression (MLR), and the advantages of using Bayesian Optimization (BO) with Gaussian Processes (GPs). This report serves as a guide to understanding the implementation and results of the study.